Cleaning up the queuing for now.  Updated the xib for a windowless application.  Temporarily fixed text color changes on drag concluded (need better solution) .

+++ ShipIt/CMPackage.m
@@ -32,7 +32,7 @@
     return [NSArray arrayWithArray: packageFiles];
 }
 
--(unsigned int)size {
+-(NSUInteger)size {
     return [packageFiles count];
 }
